refactor: native code lib/src/native -> src and 3rd_party -> third_party
